% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-json.R
\name{format_selected}
\alias{format_selected}
\title{Format the selected JSON elements}
\usage{
format_selected(json, options = NULL)
}
\arguments{
\item{json}{tsjson object.}
}
\value{
The updated tsjson object.
}
\description{
Format the selected JSON elements
}
\details{
If \code{json} does not have a selection, then all of it is formatted.
If \code{json} has an empty selection, then nothing happens.
}
\examples{
json <- parse_json(text = "{ \"a\": [1,2,3] }")
json

json |> format_selected()

json |> select("a") |> format_selected()
}
