% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert-into-selections.R
\name{insert_into_selected}
\alias{insert_into_selected}
\title{Insert a new element into the selected ones in a tsjson object}
\usage{
insert_into_selected(json, new, key = NULL, at = Inf, options = NULL)
}
\arguments{
\item{json}{tsjson object}

\item{new}{New element to insert. Will be serialized with
\code{\link[=serialize_json]{serialize_json()}}.}

\item{key}{Key of the new element, when inserting into an object.}

\item{at}{What position to insert the new element at:
\itemize{
\item \code{0}: at the beginning,
\item \code{Inf}: at the end,
\item other numbers: after the specified element,
\item a character scalar, the key after which the new element is inserted,
if that key exists, when inserting into an object. If this key does
not exist, then the new element is inserted at the end of the object.
}}
}
\value{
The modified tsjson object.
}
\description{
Insert a new element into each selected array or object.
}
\details{
It is an error trying to insert into an element that is not an array and
not an object.
}
\examples{
json <- parse_json(text = "{ \"a\": true, \"b\": [1, 2, 3] }")
json

json |> select("b") |> insert_into_selected("foo", at = 1)
}
