% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-json.R
\name{parse_json}
\alias{parse_json}
\title{Parse a JSON file or string into a tsjson object}
\usage{
parse_json(file = NULL, text = NULL, ranges = NULL, options = NULL)
}
\value{
A tsjson object.
}
\description{
Parse a JSON file or string and create a tsjson object that represents
its document. This object can then be queried and manipulated.
}
\details{
tsjson objects have \code{\link[=format.tsjson]{format()}} and
\code{\link[=print.tsjson]{print()}} methods to pretty-print them to the screen.

They can be converted to a data frame using the
\link[=tsjson-brackets]{single bracket} operator.
}
\examples{
text <- '
{
  "a": 1,
  "b": [2, 3, 4],
  "[r]": {
    "this": "setting",
    // A comment!
    "that": true
  }
}
'

# Parse the JSON, allowing comments (i.e. JSONC)
parse_json(text = text)

# Try to parse the JSON, but comments aren't allowed!
try(parse_json(text = text, options = list(allow_comments = FALSE)))

# Extract parts of the JSON
parse_json(text = text) |> select("b") |> unserialize_selected()
parse_json(text = text) |> select("[r]") |> unserialize_selected()
parse_json(text = text) |> select("[r]", "that") |> unserialize_selected()

# Use a `list()` combining strings and positional indices when
# arrays are involved
parse_json(text = text) |> select("b", 2) |> unserialize_selected()
}
\seealso{
\code{\link[=select]{select()}} to select part(s) of a tsjson object,
\code{\link[=unserialize_selected]{unserialize_selected()}} to extract the selected part(s),
\code{\link[=format_selected]{format_selected()}} to format the selected part(s),
\code{\link[=delete_selected]{delete_selected()}}, \code{\link[=insert_into_selected]{insert_into_selected()}} and
\code{\link[=update_selected]{update_selected()}} to manipulate it. \code{\link[=save_json]{save_json()}} to save
the JSON document to a file.
}
