% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-json.R
\name{save_json}
\alias{save_json}
\title{Write a tsjson object to a file}
\usage{
save_json(json, file = NULL)
}
\arguments{
\item{json}{tsjson object.}

\item{file}{File or connection to write to. Both binary and text
connections are supported. Use \code{stdout()} to output to the screen.}
}
\value{
Nothing.
}
\description{
Write a tsjson object to a file
}
\seealso{
\code{\link[=parse_json]{parse_json()}} to create a tsjson object from a JSON file or
string. \code{\link[=unserialize_selected]{unserialize_selected()}} obtain a JSON string for a tsjson
object.
}
