% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-format-jsonc.R
\name{ts_format_jsonc}
\alias{ts_format_jsonc}
\title{Format a JSONC file}
\usage{
ts_format_jsonc(file = NULL, text = NULL, options = NULL)
}
\arguments{
\item{file}{Path to a JSONC file. Use one of \code{file} or \code{text}.}

\item{text}{JSONC content as a raw vector or a character vector. Use one
of \code{file} or \code{text}.}

\item{options}{Named list of parsing and formatting options, see
\link[=tsjsonc_options]{tsjsonc options}.}
}
\value{
Nothing.
}
\description{
Format a JSONC file
}
