% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsjsonc-package.R
\name{ts_language_jsonc}
\alias{ts_language_jsonc}
\title{Tree sitter language object for JSONC}
\usage{
ts_language_jsonc()
}
\description{
Use this function with \code{\link[ts:ts_tree_new]{ts::ts_tree_new()}} to create a tree-sitter
tree for a JSONC document.
}
