% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-tree-new.R
\name{ts_tree_new.ts_language_jsonc}
\alias{ts_tree_new.ts_language_jsonc}
\alias{ts_tree_jsonc}
\title{Create a new tree-sitter tree for a JSONC document}
\usage{
\method{ts_tree_new}{ts_language_jsonc}(
  language,
  file = NULL,
  text = NULL,
  ranges = NULL,
  fail_on_parse_error = TRUE,
  options = NULL,
  ...
)
}
\arguments{
\item{language}{\eval{ts:::doc_insert("ts::ts_tree_new_param_language", "tsjsonc")}}

\item{file}{\eval{ts:::doc_insert("ts::ts_tree_new_param_file", "tsjsonc")}}

\item{text}{\eval{ts:::doc_insert("ts::ts_tree_new_param_text", "tsjsonc")}}

\item{ranges}{\eval{ts:::doc_insert("ts::ts_tree_new_param_ranges", "tsjsonc")}}

\item{fail_on_parse_error}{\eval{ts:::doc_insert("ts::ts_tree_new_param_fail_on_parse_error",
"tsjsonc")}}

\item{options}{Named list of formatting options, see
\link[=tsjsonc_options]{tsjsonc options}.}

\item{...}{Reserved for future use.}
}
\value{
\eval{ts:::doc_insert("ts::ts_tree_new_return", "tsjsonc")}
}
\description{
\eval{ts:::doc_insert("ts::ts_tree_new_description", "tsjsonc")}
}
\details{
\if{text}{This is the S3 method of the [ts::ts_tree_new()] generic.}

\eval{ts:::doc_extra()}
\eval{ts:::doc_insert("ts::ts_tree_new_details", "tsjsonc")}
}
