% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-tree-select.R
\name{ts_tree_select.tsjsonc}
\alias{ts_tree_select.tsjsonc}
\title{Select parts of a JSONC tree-sitter tree}
\usage{
\method{ts_tree_select}{tsjsonc}(tree, ..., refine = FALSE)
}
\arguments{
\item{tree}{\eval{ts:::doc_insert("ts::ts_tree_select_param_tree", "tsjsonc")}}

\item{...}{\eval{ts:::doc_insert("ts::ts_tree_select_param_dots", "tsjsonc")}}

\item{refine}{\eval{ts:::doc_insert("ts::ts_tree_select_param_refine", "tsjsonc")}}
}
\value{
\eval{ts:::doc_insert("ts::ts_tree_select_return", "tsjsonc")}
}
\description{
\eval{ts:::doc_insert("ts::ts_tree_select_description", "tsjsonc")}

This is the S3 method of the \code{\link[ts:ts_tree_select]{ts::ts_tree_select()}} generic,
for \link{tsjsonc} objects.
}
\details{
\eval{ts:::doc_insert("ts::ts_tree_select_details", "tsjsonc")}
\eval{ts:::doc_insert("tsjsonc::ts_tree_select_examples", "tsjsonc")}
\eval{ts:::doc_extra()}
}
