% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-json.R
\name{ts_tree_write}
\alias{ts_tree_write}
\title{Write a tsjsonc object to a file}
\arguments{
\item{json}{tsjsonc object.}

\item{file}{File or connection to write to. Both binary and text
connections are supported. Use \code{stdout()} to output to the screen.}
}
\value{
Nothing.
}
\description{
Write a tsjsonc object to a file
}
\seealso{
\code{\link[=ts_read_jsonc]{ts_read_jsonc()}} to create a tsjsonc object from a JSON
file or string. \code{\link[ts:ts_tree_unserialize]{ts::ts_tree_unserialize()}} obtain a JSON string for a
tsjsonc object.
}
