% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsjsonc-package.R
\docType{package}
\name{tsjsonc-package}
\alias{tsjsonc}
\alias{tsjsonc-package}
\title{tsjsonc: Edit JSON Files}
\description{
Extract and manipulate parts of JSON files without touching the formatting and comments in other parts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gaborcsardi/tsjsonc}
  \item Report bugs at \url{https://github.com/gaborcsardi/tsjsonc/issues}
}

}
\author{
\strong{Maintainer}: Gábor Csárdi \email{csardi.gabor@gmail.com}

Other contributors:
\itemize{
  \item Posit Software, PBC (\href{https://ror.org/03wc8by49}{ROR}) [copyright holder, funder]
  \item  tree-sitter-json authors [contributor]
}

}
\keyword{internal}
