% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{tsjsonc_options}
\alias{tsjsonc_options}
\title{tsjsonc options}
\description{
Options that control the behavior of tsjsonc functions.
}
\details{
\subsection{Parsing options:}{
\itemize{
\item \code{allow_empty_content}: logical, whether to allow empty JSON documents.
Default is TRUE.
\item \code{allow_comments}: logical, whether to allow comments in JSON documents.
Default is TRUE.
\item \code{allow_trailing_comma}: logical, whether to allow trailing commas in
JSON documents. Default is TRUE.
}
}

\subsection{Formatting options:}{
\itemize{
\item \code{format}: Formatting style, one of:
\itemize{
\item \code{"pretty"}: arrays and objects are formatted in multiple lines,
\item \code{"compact"}: format everything without whitespace,
\item \code{"oneline"}: format everything without newlines, but include
whitespace after commas, colons, opening brackets and braces, and
before closing brackets and braces.
Default is \verb{rpretty* }indent_width\verb{: integer, the number of spaces to use for indentation when }indent_style\code{is}"space"`. Default is
}
\enumerate{
\item 
}
\item \code{indent_style}: string, either \code{"space"} or \code{"tab"}, the type of
indentation to use. Default is space.
}
}
}
