% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unserialize-json.R
\name{unserialize_selected}
\alias{unserialize_selected}
\title{Unserialize selected elements from a tsjson object}
\usage{
unserialize_selected(json)
}
\arguments{
\item{json}{tsjson object.}
}
\value{
List of R objects, each the unserialization of a selected element
in tsjson.
}
\description{
Uses \code{\link[=unserialize_json]{unserialize_json()}} on the selected elements.
}
\details{
If \code{json} does not have a selection, then all of it is unserialized.
If \code{json} has an empty selection, then an empty list is returned.
}
\examples{
json <- parse_json(text = serialize_json(list(
  a = list(a1 = list(1,2,3), a2 = "string"),
  b = list(4, 5, 6),
  c = list(c1 = list("a", "b"))
)))
json |> select(c("b", "c")) |> unserialize_selected()
}
\seealso{
\code{\link[=unserialize_json]{unserialize_json()}} to unserialize a JSON document from a
file or string. \code{\link[=serialize_json]{serialize_json()}} to create JSON from R objects.
}
