% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-tree-read-jsonc.R
\name{ts_parse_jsonc}
\alias{ts_parse_jsonc}
\alias{ts_read_jsonc}
\title{Parse a JSON file or string into a tsjsonc object}
\usage{
ts_parse_jsonc(text, ranges = NULL, fail_on_parse_error = TRUE, options = NULL)

ts_read_jsonc(file, ranges = NULL, fail_on_parse_error = TRUE, options = NULL)
}
\arguments{
\item{text}{String. Use either \code{file} or \code{text}, but not both.}

\item{ranges}{Can be used to parse part(s) of the input. It must be a
data frame with integer columns \code{start_row}, \code{start_col}, \code{end_row},
\code{end_col}, \code{start_byte}, \code{end_byte}, in this order.}

\item{fail_on_parse_error}{Logical, whether to error if there are
parse errors in the document. Default is \code{TRUE}.}

\item{options}{Named list of parsing options, see
\link[=tsjsonc_options]{tsjsonc options}.}

\item{file}{Path of a file. Use either \code{file} or \code{text}.}
}
\value{
A tsjsonc object.
}
\description{
Parse a JSON file or string and create a tsjsonc object that
represents its document. This object can then be queried and
manipulated.
}
\details{
tsjsonc objects have \code{format()} and \code{print()} methods to pretty-print
them to the screen.

They can be converted to a data frame using the
\link[ts:ts_tree-brackets]{single bracket} operator.
}
\examples{
text <- '
{
  "a": 1,
  "b": [2, 3, 4],
  "[r]": {
    "this": "setting",
    // A comment!
    "that": true
  }
}
'

# Parse the JSON, allowing comments (i.e. JSONC)
ts_parse_jsonc(text)

# Try to parse the JSON, but comments aren't allowed!
try(ts_parse_jsonc(text, options = list(allow_comments = FALSE)))

# Extract parts of the JSON
ts_parse_jsonc(text) |>
  ts_tree_select("b") |>
  ts_tree_unserialize()
ts_parse_jsonc(text) |>
  ts_tree_select("[r]") |>
  ts_tree_unserialize()
ts_parse_jsonc(text) |>
  ts_tree_select("[r]", "that") |>
  ts_tree_unserialize()

# Use a `list()` combining strings and positional indices when
# arrays are involved
ts_parse_jsonc(text) |>
  ts_tree_select("b", 2) |>
  ts_tree_unserialize()
}
\seealso{
\code{\link[=ts_tree_select]{ts_tree_select()}} to select part(s) of a tsjsonc
object, \code{\link[ts:ts_tree_unserialize]{ts::ts_tree_unserialize()}} to extract the selected part(s),
\code{\link[ts:ts_tree_format]{ts::ts_tree_format()}} to format the selected part(s),
\code{\link[ts:ts_tree_delete]{ts::ts_tree_delete()}}, \code{\link[ts:ts_tree_insert]{ts::ts_tree_insert()}} and
\code{\link[ts:ts_tree_update]{ts::ts_tree_update()}} to manipulate it. \code{\link[=ts_tree_write]{ts_tree_write()}} to save
the JSON document to a file.
}
