% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{ts_tree_delete.ts_tree_jsonc}
\alias{ts_tree_delete.ts_tree_jsonc}
\title{Delete selected elements from a tsjsonc object}
\usage{
\method{ts_tree_delete}{ts_tree_jsonc}(tree, ...)
}
\arguments{
\item{tree}{tsjsonc object.}

\item{...}{Reserved for future use.}
}
\value{
Modified tsjsonc object.
}
\description{
The formatting of the rest of JSON document is kept as is. Comments
appearing inside the deleted elements are also deleted. Other comments
are left as is.
}
\details{
If \code{tree} has no selection then the the whole document is deleted.
If \code{tree} has an empty selection, then nothing is delted.
}
\examples{
library(ts)
tree <- ts_parse_jsonc("{ \"a\": //comment\ntrue, \"b\": [1, 2, 3] }")
tree

tree |> ts_tree_select("a")
tree |> ts_tree_select("a") |> ts_tree_delete()
}
