% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-tree-select.R
\name{ts_tree_select.tsjsonc}
\alias{ts_tree_select.tsjsonc}
\title{Select parts of a JSONC tree-sitter tree}
\usage{
\method{ts_tree_select}{tsjsonc}(tree, ..., refine = FALSE)
}
\arguments{
\item{tree}{\eval{tsjsonc:::docs("ts_tree_select", "param-tree")}}

\item{...}{\eval{tsjsonc:::docs("ts_tree_select", "param-...")}}

\item{refine}{\eval{tsjsonc:::docs("ts_tree_select", "param-refine")}}
}
\value{
\eval{tsjsonc:::docs("ts_tree_select", "return")}
}
\description{
\eval{tsjsonc:::docs("ts_tree_select", "description")}
}
\details{
\eval{tsjsonc:::docs("ts_tree_select", "details")}
}
\examples{
# hardcoded examples
}
