% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unserialize-json.R
\name{ts_unserialize_jsonc}
\alias{ts_unserialize_jsonc}
\title{Unserialize a JSON file or string into an R object}
\usage{
ts_unserialize_jsonc(file = NULL, text = NULL, ranges = NULL, options = NULL)
}
\arguments{
\item{file}{Path of a file. Use either \code{file} or \code{text}, but not both.}

\item{text}{String. Use either \code{file} or \code{text}, but not both.}

\item{ranges}{Can be used to parse part(s) of the input. It must be a
data frame with integer columns \code{start_row}, \code{start_col}, \code{end_row},
\code{end_col}, \code{start_byte}, \code{end_byte}, in this order.}

\item{options}{Named list of parsing options, see
\link[=tsjsonc_options]{tsjsonc options}.}
}
\value{
R object.
}
\description{
The purpose of this function is to convert a JSON file or string into
an R object reliably.
}
\details{
See examples below on how the different JSON elements are mapped to
R objects.
}
\examples{
library(ts)
# null -> NULL
ts_unserialize_jsonc(text = "null")

# true, false -> TRUE, FALSE
ts_unserialize_jsonc(text = "true")
ts_unserialize_jsonc(text = "false")

# string -> character scalar
ts_unserialize_jsonc(text = "\"string with escapes: \\\\b \\\\ud020\"")

# number -> double scalar
ts_unserialize_jsonc(text = "42.25")

# array -> unnamed list
ts_unserialize_jsonc(text = "[1, 2, 3]")

# object -> named list
ts_unserialize_jsonc(text = "{\"a\": 1, \"b\": 2 }")
}
\seealso{
\code{\link[=ts_serialize_jsonc]{ts_serialize_jsonc()}} for the opposite, \code{\link[ts:ts_tree_select]{ts::ts_tree_select()}} and
\code{\link[ts:ts_tree_unserialize]{ts::ts_tree_unserialize()}} to unserialize part(s) of a JSON document.
\code{\link[ts:ts_tree_new]{ts::ts_tree_new()}} to load a JSON document and then manipulate it.
}
